+++ CHANGELOG.md
@@ -1,10 +1,16 @@
-#0.0.4
+# 1.0.0
+* Updated to React 17.
+* Renders to a `React.Fragment` now. Does not create an additional div.
+* Support for custom renders.
+* `useLocaleString` deprecated.
+
+# 0.0.4
 Fix incorrect numbers being drawn when the screen was inactive. (thanks! @migreva)
 
-#0.0.3
+# 0.0.3
 Added `delayValue` prop to delay the start of the animation.
 
-#0.0.2
+# 0.0.2
 Added `userLocaleString` prop. Thanks @adjohu.
 
 #0.0.1